+++ debugger/src/memory_breakpoints.rs
@@ -6,8 +6,8 @@ mod breakpoints_test;
 
 #[derive(Default)]
 pub struct MemoryBreakpoints {
-    breakpoints: Vec<usize>,
-    map: HashMap<usize, u8>,
+    breakpoints: Vec<u32>,
+    map: HashMap<u32, u8>,
 }
 
 // a list of addresses for the debugger to break on when memory content changes
@@ -19,7 +19,7 @@ impl MemoryBreakpoints {
         }
     }
 
-    pub fn add(&mut self, bp: usize) -> Option<usize> {
+    pub fn add(&mut self, bp: u32) -> Option<u32> {
         if self.breakpoints.iter().find(|&&x|x == bp).is_none() {
             self.breakpoints.push(bp);
             Some(bp)
@@ -28,7 +28,7 @@ impl MemoryBreakpoints {
         }
     }
 
-    pub fn remove(&mut self, bp: usize) -> Option<usize> {
+    pub fn remove(&mut self, bp: u32) -> Option<u32> {
         // TODO later: simplify when https://github.com/rust-lang/rust/issues/40062 is stable
         match self.breakpoints.iter().position(|x| *x == bp) {
             Some(pos) => {
@@ -40,7 +40,7 @@ impl MemoryBreakpoints {
     }
 
     // returns a Vec with breakpoints sorted ascending
-    pub fn get(&self) -> Vec<usize> {
+    pub fn get(&self) -> Vec<u32> {
         let mut sorted = self.breakpoints.clone();
         sorted.sort();
         sorted
@@ -51,7 +51,7 @@ impl MemoryBreakpoints {
     }
 
     // returns true memory value has changed since last check
-    pub fn has_changed(&mut self, address: usize, val: u8) -> bool {
+    pub fn has_changed(&mut self, address: u32, val: u8) -> bool {
         if !self.map.contains_key(&address) {
             self.map.insert(address, val);
             return false;
